Common, PWGEM/Dilepton: remove unnecessary tasks
