+++ Common/TableProducer/CMakeLists.txt
@@ -150,8 +150,3 @@ o2physics_add_dpl_workflow(muon-realignment
                     PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase O2::DetectorsCommonDataFormats O2::MathUtils O2::MCHTracking O2::DataFormatsMCH O2::GlobalTracking O2::MCHBase O2::MCHGeometryTransformer O2::CommonUtils
                     COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(fwdtrack-propagation
-                    SOURCES fwdtrackPropagation.cxx
-                    PUBLIC_LINK_LIBRARIES O2Physics::AnalysisCore O2::GlobalTracking
-                    COMPONENT_NAME Analysis)
-
